Rename placement helper to free slot
